+++ api/routes/talkLogs.js
